+++ sdk/src/DocuSign.Click/Properties/AssemblyInfo.cs
@@ -22,5 +22,5 @@ using System.Runtime.InteropServices;
 // [assembly: AssemblyVersion("1.0.*")]
 internal class AssemblyInformation
 {
-    public const string AssemblyInformationalVersion = "1.0.0-beta";
+    public const string AssemblyInformationalVersion = "1.0.0";
 }