+++ sdk/CHANGELOG.md
@@ -1,6 +1,10 @@
 # DocuSign Click C# Client Changelog
 See [DocuSign Support Center](https://support.docusign.com/en/releasenotes/) for Product Release Notes.
 
+## [v1.0.0] - DocuSign Click API v1 - 02/11/2021
+### Changed
+- First GA version of Click API, supports DocuSign Click v1-20.4.02.
+
 ## [v1.0.0-beta] - DocuSign Click API v1 - 12/17/2020
 ### Changed
 - First Beta version of Click API.